# docs/swagger.yaml
basePath: /v1
definitions:
...
  entity.User:
    properties:
      avatarUrl:
        type: string
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      email:
        example: john@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      isActive:
        example: true
        type: boolean
      nativeLanguageId:
        type: integer
      premiumUntil:
        type: string
      role:
        example: user
        type: string
      targetLanguageId:
        type: integer
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      username:
        example: johndoe
        type: string
    type: object
...
  github_com_evrone_go-clean-template_internal_controller_restapi_v1_admin_request.Login:
    properties:
      idToken:
        example: firebase-id-token
        type: string
    required:
    - idToken
    type: object
...
  github_com_evrone_go-clean-template_internal_controller_restapi_v1_admin_response.Login:
    properties:
      authenticated:
        example: true
        type: boolean
      user:
        $ref: '#/definitions/entity.User'
    type: object
...
  internal_controller_restapi_v1_admin.errorDoc:
    properties:
      error:
        example: message
        type: string
    type: object
...
  response.OnboardingLanguage:
    properties:
      code:
        example: en
        type: string
      flag:
        example: "\U0001F1EC\U0001F1E7"
        type: string
      id:
        example: 1
        type: integer
      name:
        example: English
        type: string
    type: object
...
  response.UserProfile:
    properties:
      avatarUrl:
        example: https://example.com/avatar.jpg
        type: string
      email:
        example: john@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      nativeLanguage:
        $ref: '#/definitions/response.OnboardingLanguage'
      targetLanguage:
        $ref: '#/definitions/response.OnboardingLanguage'
      username:
        example: johndoe
        type: string
    type: object
...
  v1.Error:
    properties:
      error:
        example: message
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: User API with Firebase authentication and role-based access control
  title: Go Clean Template API
  version: "1.0"
paths:
  /admin/auth/login:
    post:
      consumes:
      - application/json
      description: Verify a Firebase ID token and require the local user to have the
        admin role
      operationId: admin-login
      parameters:
      - description: Firebase credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_admin_request.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_admin_response.Login'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
      summary: Admin login
      tags:
      - admin
...
  /admin/me:
    get:
      description: Return the authenticated administrator profile
      operationId: admin-me
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_restapi_v1_admin.errorDoc'
      security:
      - BearerAuth: []
      summary: Get current admin
      tags:
      - admin
...
  /user/profile:
    get:
      description: Get current user profile
      operationId: profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserProfile'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Get profile
      tags:
      - user
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"